# Optimize Set is_disjoint