@Amanieu 2024-06-10
Just add `#[cfg(feature = "raw")]` to these types. The CI failure is probably due to lint changes in nightly rustc.